requirements.txt
consul
requests
schedule
pyyaml